Fixed automerge issue and renamed bullet.py to projectile.py (#56)
